Update MeowGame.java
Should be cleaner and holds less details that it shouldnt have. Will remove MouseListener after I've created all the area specific Rectangles. Honestly it doesn't work as intended at the moment but it was never meant to make it to the end game anyways, I just have to draw the rectangles here on the screen n play with it.