Add files via upload
Holds the information for image creation, the Z is meant more for scrolling but at the moment I don't plan to implement scrolling. This should make image creation uniform so there should be no error with image creation unless its name was spelt wrong or it had nothing to do with Image creation in the first place.